@@ -8,20 +8,16 @@
     "build": "webpack --mode production"
   },
   "devDependencies": {
-    "html-webpack-plugin": "^3.2.0",
-    "prettier": "^1.16.4",
-    "ts-loader": "^5.3.3",
-    "typescript": "^3.3.3333",
-    "webpack": "^4.29.5",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.2.0"
+    "@vitejs/plugin-react": "^5.0.1",
+    "prettier": "latest",
+    "typescript": "^5.9.2",
+    "vite": "^7.1.3"
   },
   "dependencies": {
-    "@emotion/core": "^10.0.7",
-    "@emotion/styled": "^10.0.7",
-    "@types/react": "^16.8.4",
-    "@types/react-dom": "^16.8.2",
-    "react": "^16.8.3",
-    "react-dom": "^16.8.3"
+    "@emotion/styled": "11",
+    "@types/react": "18",
+    "@types/react-dom": "18",
+    "react": "18",
+    "react-dom": "18"
   }
 }
